Move Vulnerabilities from workload to platform
