+++ src/cbc_sdk/platform/__init__.py
@@ -18,3 +18,5 @@ from cbc_sdk.platform.reputation import ReputationOverride
 from cbc_sdk.platform.grants import Grant
 
 from cbc_sdk.platform.users import User
+
+from cbc_sdk.platform.vulnerability_assessment import Vulnerability
