+++ src/cbc_sdk/workload/__init__.py
@@ -1,3 +1,5 @@
 from cbc_sdk.workload.vm_workloads_search import ComputeResource
-from cbc_sdk.workload.vulnerability_assessment import Vulnerability
 from cbc_sdk.workload.sensor_lifecycle import SensorKit
+
+# Maintain link for easier migration
+from cbc_sdk.platform.vulnerability_assessment import Vulnerability
