+++ src/tests/uat/vulnerability_assessment_uat.py
@@ -37,8 +37,7 @@ import requests
 
 # Internal library imports
 from cbc_sdk.helpers import build_cli_parser, get_cb_cloud_object
-from cbc_sdk.workload import Vulnerability
-from cbc_sdk.platform import Device
+from cbc_sdk.platform import Device, Vulnerability
 from cbc_sdk.errors import MoreThanOneResultError
