+++ src/tests/unit/fixtures/platform/mock_vulnerabilities.py
