+++ src/tests/unit/platform/test_vulnerability_assessment.py
@@ -18,9 +18,8 @@ import logging
 from cbc_sdk.rest_api import CBCloudAPI
 from cbc_sdk.errors import ApiError, ObjectNotFoundError, MoreThanOneResultError
 from tests.unit.fixtures.CBCSDKMock import CBCSDKMock
-from cbc_sdk.workload import Vulnerability
-from cbc_sdk.platform import Device
-from tests.unit.fixtures.workload.mock_vulnerabilities import (GET_VULNERABILITY_SUMMARY_ORG_LEVEL,
+from cbc_sdk.platform import Device, Vulnerability
+from tests.unit.fixtures.platform.mock_vulnerabilities import (GET_VULNERABILITY_SUMMARY_ORG_LEVEL,
                                                                GET_VULNERABILITY_SUMMARY_ORG_LEVEL_PER_SEVERITY,
                                                                GET_ASSET_VIEW_VUL_RESP,
                                                                GET_VULNERABILITY_RESP,
